add Blue Star award type
Coming back to Ruby I am glad to see the language is adopting more and more aspects of functional programming. I was happy to learn they adopted the pipeline `|>` operator from F#, and was just as happy to see they removed it rather doing it in a sub-standard style.

The project requirements said to keep in mind the need to work with other legacy elements of the system. As such, I stuck with creating new data rather than mutating data. I also went with smaller, composable function separate from the state. I considered having one ancestor award class with several descendants. I decided against that due to not knowing about other system integration requirements. However, the code is composable and the functions discreet enough that moving them to instance methods on classes would be trivial. Also, having all the variations of the behavior in one location favors safety and reasoning about the code.